Use float instead of double

+++ cpp/Sample.h
@@ -11,12 +11,12 @@ public:
 
     bool run();
 
-    double r2;
+    float r2;
 };
 
 inline bool Sample::run() {
     for (int j = 0; j < d; j++) {
-        double x_j = get_random_0_to_1();
+        float x_j = get_random_0_to_1();
         r2 += pow(x_j, 2);
         if (r2 > 1) return false;
     }
